requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
spacy==3.7.2
feedparser==6.0.10
pydantic==2.5.0
google-generativeai==0.3.0
openai==1.3.0
python-dateutil==2.8.2
tenacity==8.2.3
